Split Telegram channel tests by responsibility
Move the flat Telegram channel test module into focused child
modules for inbound handling, DTO parsing, polling, status mapping,
attachment extraction, downloads, sending, state constants, and webhook
helpers.

Keep test names and assertions intact while narrowing imports and
exposing only the crate-private helpers needed by the moved tests.
