+++ channels-src/telegram/src/tests/attachments.rs
@@ -1,0 +1,187 @@
+use crate::attachments::extract_attachments;
+use crate::types::TelegramMessage;
+
+// === Attachment extraction fixture tests ===
+
+#[test]
+fn test_extract_attachments_photo() {
+    let json = r#"{
+        "message_id": 1,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "caption": "What is this?",
+        "photo": [
+            {"file_id": "small_id", "file_unique_id": "s1", "width": 90, "height": 90, "file_size": 1234},
+            {"file_id": "large_id", "file_unique_id": "l1", "width": 800, "height": 600, "file_size": 54321}
+        ]
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    assert_eq!(attachments.len(), 1);
+    assert_eq!(attachments[0].id, "large_id"); // Largest photo
+    assert_eq!(attachments[0].mime_type, "image/jpeg");
+    assert_eq!(attachments[0].size_bytes, Some(54321));
+    assert!(attachments[0]
+        .source_url
+        .as_ref()
+        .unwrap()
+        .contains("large_id"));
+}
+
+#[test]
+fn test_extract_attachments_document() {
+    let json = r#"{
+        "message_id": 2,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "document": {
+            "file_id": "doc_abc",
+            "file_unique_id": "d1",
+            "file_name": "report.pdf",
+            "mime_type": "application/pdf",
+            "file_size": 102400
+        },
+        "caption": "Here is the report"
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    assert_eq!(attachments.len(), 1);
+    assert_eq!(attachments[0].id, "doc_abc");
+    assert_eq!(attachments[0].mime_type, "application/pdf");
+    assert_eq!(attachments[0].filename, Some("report.pdf".to_string()));
+    assert_eq!(attachments[0].size_bytes, Some(102400));
+}
+
+#[test]
+fn test_extract_attachments_voice() {
+    let json = r#"{
+        "message_id": 3,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "voice": {
+            "file_id": "voice_xyz",
+            "file_unique_id": "v1",
+            "duration": 5,
+            "mime_type": "audio/ogg",
+            "file_size": 9000
+        }
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    assert_eq!(attachments.len(), 1);
+    assert_eq!(attachments[0].id, "voice_xyz");
+    assert_eq!(attachments[0].mime_type, "audio/ogg");
+    assert_eq!(
+        attachments[0].filename.as_deref(),
+        Some("voice_voice_xyz.ogg")
+    );
+    assert!(attachments[0].extras_json.contains("\"duration_secs\":5"));
+}
+
+#[test]
+fn test_extract_attachments_video() {
+    let json = r#"{
+        "message_id": 4,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "video": {
+            "file_id": "vid_1",
+            "file_unique_id": "vv1",
+            "file_name": "clip.mp4",
+            "mime_type": "video/mp4",
+            "file_size": 5000000
+        },
+        "caption": "Check this out"
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    assert_eq!(attachments.len(), 1);
+    assert_eq!(attachments[0].id, "vid_1");
+    assert_eq!(attachments[0].mime_type, "video/mp4");
+    assert_eq!(attachments[0].filename, Some("clip.mp4".to_string()));
+}
+
+#[test]
+fn test_extract_attachments_audio() {
+    let json = r#"{
+        "message_id": 5,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "audio": {
+            "file_id": "audio_1",
+            "file_unique_id": "a1",
+            "file_name": "song.mp3",
+            "mime_type": "audio/mpeg",
+            "file_size": 3000000
+        }
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    assert_eq!(attachments.len(), 1);
+    assert_eq!(attachments[0].id, "audio_1");
+    assert_eq!(attachments[0].mime_type, "audio/mpeg");
+    assert_eq!(attachments[0].filename, Some("song.mp3".to_string()));
+}
+
+#[test]
+fn test_extract_attachments_sticker() {
+    let json = r#"{
+        "message_id": 6,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "sticker": {
+            "file_id": "sticker_1",
+            "file_unique_id": "st1",
+            "type": "regular",
+            "file_size": 20000
+        }
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    assert_eq!(attachments.len(), 1);
+    assert_eq!(attachments[0].id, "sticker_1");
+    assert_eq!(attachments[0].mime_type, "image/webp");
+}
+
+#[test]
+fn test_extract_attachments_text_only_empty() {
+    let json = r#"{
+        "message_id": 7,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "text": "Hello"
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    assert!(attachments.is_empty());
+}
+
+#[test]
+fn test_extract_attachments_multiple_types() {
+    let json = r#"{
+        "message_id": 8,
+        "from": {"id": 1, "is_bot": false, "first_name": "A"},
+        "chat": {"id": 1, "type": "private"},
+        "photo": [
+            {"file_id": "photo_1", "file_unique_id": "p1", "width": 100, "height": 100}
+        ],
+        "document": {
+            "file_id": "doc_1",
+            "file_unique_id": "d1", 
+            "file_name": "file.txt",
+            "mime_type": "text/plain"
+        }
+    }"#;
+    let msg: TelegramMessage = serde_json::from_str(json).unwrap();
+    let attachments = extract_attachments(&msg);
+
+    // Both photo and document should be extracted
+    assert_eq!(attachments.len(), 2);
+}
