+++ channels-src/telegram/src/tests/mod.rs
@@ -1,0 +1,9 @@
+mod attachments;
+mod downloads;
+mod inbound;
+mod polling;
+mod send;
+mod state;
+mod status;
+mod types;
+mod webhook;
